requirements.txt
requests
psycopg2-binary
python-dotenv
pytz
azure-functions
requests
